Split initiation test over multiple files
A feature of tale-maker will be to freely concatenate files in an order
agnostic way. Each end-of-file will also be an implicit end-of-block,
but otherwise the files distinctions will be ignored by the parser.
Errors will be thrown on compile if there are redundant or missing
declarations.

One interesting feature of the file flexibility, is I can have one file
for Taelmoor configs (especially using the piece codes as refs) and
another file for the tale-player (CLI text adventure) configs (english
word refs). You will be able to include one file or the other during
compilation.