Refactor lexer to advance monotonically
Who doesn't love an excuse to use the term "monotonic"?

The big issue was that the previous version of the lexer had no way to track line numbers. The somewhat smaller issue was that the lexer jumped around a bit which no doubt introduced some inefficiencies.

This new implementation only ever advances forward, and does so in a
centralized location where I will be able to track line numbers later.
It's also generally a simpler implementation which should generally be
easier to maintain and add features to.