Replace recursive no-op with loop
It is only really possible to have to recurse once with the current
implementation (scanWhileText compresses empty lines), but when I
implement actions it will be possible to repeatedly no-op with a bunch
of empty lines in the middle of the action. *Probably* unlikely anyone
blows the stack that way, but it is easy enough to swap out the
recursive no-ops with a loop, so let's just do that.