Handle ARG with no padding after
Whoops. It is totally valid to place a header end character directly
after an ARG. Added an isAnyOf helper so I can test for both that and
whitespace. It's possible I should go the other way and create a
limited set of valid ARG characters that does not include block header
characters... but I am not prepared to make that design decision yet.

For now at least, any character that I can contextually disambiguate
from other syntax is a valid ARG character.