Lex actions
Actions are directives which can be placed anywhere, which introduces
the possibility that there is more than one Text token within a single
block. This added some extra complication to the empty-line trimming
logic as we only want to trim empty lines at the start and end of a
block and want to retain all whitespace that surrounds an action.